# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"createPaymentRequest": {
				"properties": {
					"member_id": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"transaction_id": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"PAID",
							"UNPAID"
						]
					}
				},
				"required": [
					"member_id",
					"amount",
					"transaction_id",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getPaymentsRequest": {
				"properties": {
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageIndex": {
						"type": "number",
						"format": "double"
					},
					"searchString": {
						"type": "string"
					},
					"sortBy": {
						"type": "string"
					},
					"sortOn": {
						"type": "string",
						"enum": [
							"ASC",
							"DESC"
						]
					}
				},
				"required": [
					"pageSize",
					"pageIndex"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetMemberRequest": {
				"properties": {
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageIndex": {
						"type": "number",
						"format": "double"
					},
					"searchString": {
						"type": "string"
					},
					"sortBy": {
						"type": "string"
					},
					"sortOn": {
						"type": "string",
						"enum": [
							"ASC",
							"DESC"
						]
					}
				},
				"required": [
					"pageSize",
					"pageIndex"
				],
				"type": "object",
				"additionalProperties": false
			},
			"createMemberRequest": {
				"properties": {
					"member_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone_no": {
						"type": "string"
					}
				},
				"required": [
					"member_name",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "funtrip",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/payment/{id}": {
			"get": {
				"operationId": "GetPaymentByMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createPaymentRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payment": {
			"post": {
				"operationId": "CreatePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createPaymentRequest"
							}
						}
					}
				}
			}
		},
		"/payment/list": {
			"post": {
				"operationId": "GetPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/getPaymentsRequest"
							}
						}
					}
				}
			}
		},
		"/member/{id}": {
			"get": {
				"operationId": "GetMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createMemberRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/member/list": {
			"post": {
				"operationId": "GetMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetMemberRequest"
							}
						}
					}
				}
			}
		},
		"/member": {
			"post": {
				"operationId": "CreateMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"additionalProperties": false,
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Member"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createMemberRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}